Remove the old BlockPulller
